Moves JSONData to JSONFunctionality
